feat: hierarchical rate limiting support

+++ src/service/promql/search/cache/mod.rs
@@ -226,7 +226,7 @@ pub async fn set(
 ) -> Result<()> {
     // check time range, if over ZO_MAX_FILE_RETENTION_TIME, return
     let cfg = get_config();
-    let max_ts = now_micros() - second_micros(cfg.limit.cache_delay_secs as i64);
+    let max_ts = now_micros() - second_micros(cfg.limit.cache_delay_secs);
     let new_end = if end > max_ts { max_ts } else { end };
     if range_values.is_empty() || start >= max_ts || new_end <= start + step {
         // all of the data in retention time, no need to store
